requirements.txt
Django~=5.1.3
djangorestframework~=3.15.2
django-simple-history~=3.7.0
django-debug-toolbar~=4.4.6
drf-spectacular~=0.27.2
djangorestframework-simplejwt~=5.3.1